tests/wpt/tests/css/css-display/reading-order/reading-order-items-computed.html

<!DOCTYPE html>
<title>CSS Display: getComputedStyle().readingOrderItems</title>
<link rel="help" href="https://drafts.csswg.org/css-display-4/#reading-order-items">
<link rel="author" title="Di Zhang" href="mailto:dizhangg@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<body>
<div id="target"></div>
<script>
test_computed_value('reading-order-items', 'normal');
test_computed_value('reading-order-items', 'flex-visual');
test_computed_value('reading-order-items', 'flex-flow');
test_computed_value('reading-order-items', 'grid-rows');
test_computed_value('reading-order-items', 'grid-columns');
test_computed_value('reading-order-items', 'grid-order');
</script>
</body>
